Updated for Python 3. Used 2to3, fixed a few byte/str issues. Py2 implicit float division with int().

+++ src/python/deNovoQualityScore/denovo.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python
 #
 # Strelka - Small Variant Caller
 # Copyright (c) 2009-2018 Illumina, Inc.
@@ -18,7 +18,7 @@
 #
 #
 
-from __future__ import print_function
+
 
 import argparse
 import csv
@@ -37,9 +37,9 @@ import cyvcf2 as vcf
 import numpy as np
 
 # check for python 2.7
-if sys.version_info[:2] != (2, 7):
-    msg = "Python 2.7 is required for execution."  # pragma: no cover
-    raise ImportError(msg)  # pragma: no cover
+#if sys.version_info[:2] != (2, 7):
+#    msg = "Python 2.7 is required for execution."  # pragma: no cover
+#    raise ImportError(msg)  # pragma: no cover
 
 # constants
 
@@ -93,6 +93,8 @@ _indel_mu_scale = 20.0
 
 
 # 4) general
+
+python_exec = '/usr/bin/python'
 
 _indel_gt_types = {
     (True, True): 'RR',
@@ -206,7 +208,7 @@ def check_pedigree_input(pedigree, samples):
             raise ValueError(msg)
 
     # check if all pedigree sample names are referenced in the header
-    for k, v in pedigree.iteritems():
+    for k, v in pedigree.items():
         if v not in samples:
             msg = "Sample '{1}' for {0} missing in the input VCF.".format(k, v)
             raise ValueError(msg)
@@ -372,8 +374,8 @@ def full_genotype_mapping(gtsu_idx):
     """Map trio sample genotypes to DNG indexing schema"""
 
     n = len(gtsu_idx)
-    full_idx = itertools.product(xrange(n), xrange(n), xrange(n))
-    idx_gtsu = {v: k for k, v in gtsu_idx.iteritems()}  # index->GT mapping
+    full_idx = itertools.product(range(n), range(n), range(n))
+    idx_gtsu = {v: k for k, v in gtsu_idx.items()}  # index->GT mapping
     full_gt = [
         "/".join([paste(idx_gtsu[k]), paste(idx_gtsu[j]), paste(idx_gtsu[i])])
         for i, j, k in full_idx
@@ -475,9 +477,10 @@ def read_prior(path, vartype):
             'names': col_names,
             'formats': col_formats
         })
-    prior = dict(zip(col_names, prior))
+    prior = dict(list(zip(col_names, prior)))
 
     prior['denovo_flag'].astype(bool)
+    prior['gt'] = prior['gt'].astype(str)
 
     if vartype.lower() in 'snv':
         pp = prior['transmission_prob']
@@ -503,10 +506,10 @@ def reorder_prior(prior, full_gt_idx):
     """Reorder imported prior to match the DNG genotype order"""
 
     p_gt = {g: i for i, g in enumerate(prior['gt'])}
-    idx_full_gt = {v: k for k, v in full_gt_idx.iteritems()}  # revert mapping
-    new_prior_ord = [p_gt[idx_full_gt[i]] for i in xrange(len(idx_full_gt))]
+    idx_full_gt = {v: k for k, v in full_gt_idx.items()}  # revert mapping
+    new_prior_ord = [p_gt[idx_full_gt[i]] for i in range(len(idx_full_gt))]
 
-    prior = {k: np.take(v, new_prior_ord, axis=0) for k, v in prior.iteritems()}
+    prior = {k: np.take(v, new_prior_ord, axis=0) for k, v in prior.items()}
 
     return prior
 
@@ -542,7 +545,7 @@ def allele_indices(alleles, genotype_index):
     """Return ordered indices for allele combinations for PL field"""
 
     indices = [genotype_index[alleles[j], alleles[i]]
-               for i in xrange(len(alleles)) for j in xrange(i + 1)]
+               for i in range(len(alleles)) for j in range(i + 1)]
 
     return indices
 
@@ -558,9 +561,9 @@ def build_PL_indices_lookup(alleles, genotype_index, ref=None):
             alleles, repeat=3),
         itertools.product(
             alleles, repeat=4))
-    alleles_comb = itertools.ifilter(
+    alleles_comb = filter(
         lambda x: len(set(x)) == min(len(x), n_uniq_alleles), alleles_comb)
-    alleles_comb = itertools.ifilter(lambda x: ref not in x[1:], alleles_comb)
+    alleles_comb = filter(lambda x: ref not in x[1:], alleles_comb)
     alleles_lookup = {
         alleles: allele_indices(alleles, genotype_index)
         for alleles in alleles_comb
@@ -800,7 +803,7 @@ def calculate_dng_DQ_indel(variant, prior, param):
 
     try:
         pl = get_PL_field(variant)
-        alleles = tuple(['R'] + ['V' for i in xrange(len(variant.ALT))])
+        alleles = tuple(['R'] + ['V' for i in range(len(variant.ALT))])
         alleles_idx = get_PL_indices_indel(alleles)
         C = map_sample_pl_indel(pl[sample_index[0]], alleles_idx)
         M = map_sample_pl_indel(pl[sample_index[1]], alleles_idx)
@@ -846,7 +849,7 @@ def calculate_dng_DQ_sv(variant, prior, param):
 
     try:
         pl = get_PL_field(variant)
-        alleles = tuple(['R'] + ['V' for i in xrange(len(variant.ALT))])
+        alleles = tuple(['R'] + ['V' for i in range(len(variant.ALT))])
         alleles_idx = get_PL_indices_indel(alleles)
         C = map_sample_pl_indel(pl[sample_index[0]], alleles_idx)
         M = map_sample_pl_indel(pl[sample_index[1]], alleles_idx)
@@ -897,7 +900,7 @@ def add_DQ_score_to_variant(variant, param, score,
         # if variant has no DQ format field yet
         # append the DQ score to the proband and 'missing' to all other samples
         fields[8] += ':' + _field_name
-        for i in xrange(param['n_samples']):
+        for i in range(param['n_samples']):
             field_value, field_index = \
                 select_sample_score(score, i, sample_index_child, decimals, null_value)
             fields[field_index] = fields[field_index] + ':' + field_value
@@ -1011,7 +1014,7 @@ def import_bed_regions(path, fieldnames=('contig', 'start', 'end', 'annotation')
                 coords = (int(entry['start']), int(entry['end']))
                 regions.setdefault(contig, list()).append(coords)
         # convert values from list to tuple
-        regions = {contig: tuple(coords) for contig, coords in regions.iteritems()}
+        regions = {contig: tuple(coords) for contig, coords in regions.items()}
     except (IOError, BaseException):
         regions = None
 
@@ -1067,7 +1070,7 @@ if __name__ == '__main__':
     parser.add_argument(
         '--model',
         required=False,
-        choices=_models.keys(),
+        choices=list(_models.keys()),
         default=_default_model_name,
         help=argparse.SUPPRESS)
 
@@ -1156,14 +1159,16 @@ if __name__ == '__main__':
     }
 
     if args.parallel:
-        cmd_header = "{} {} --proband {} --mother {} --father {} --vcf-output {} {}".format(
-            'python -E -s', __file__, args.proband, args.mother, args.father, 'header', copy.copy(args.in_vcf))
-        denovo_cmd = "'{} {} --proband {} --mother {} --father {} --vcf-output {} -'".format(
-            'python -E -s', __file__, args.proband, args.mother, args.father, 'body')
+        cmd_header = "{} {} {} --proband {} --mother {} --father {} --vcf-output {} {}".format(
+            python_exec, '-E -s', __file__, args.proband, args.mother, args.father, 'header', copy.copy(args.in_vcf))
+        denovo_cmd = "'{} {} {} --proband {} --mother {} --father {} --vcf-output {} -'".format(
+            python_exec, '-E -s', __file__, args.proband, args.mother, args.father, 'body')
         cmd_unzip = "gunzip -cdf {}".format(
             args.in_vcf)
         cmd_body = "parallel -j {} --header '(\#.*\n)*' --keep-order --pipe --no-notice {}".format(
             args.ncores, denovo_cmd)
+
+        p2 = None
 
         try:
             output_file = select_output(args.out_vcf, 'w')
@@ -1175,7 +1180,8 @@ if __name__ == '__main__':
             p2.wait()
             # handle premature closing of stdout
         except (KeyboardInterrupt, BaseException):
-            p2.terminate()
+            if p2 is not None:
+               p2.terminate()
 
     else:
         denovo(args.in_vcf, pedigree, param, args.out_vcf)
